# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Astral Core v7 API
  version: 7.0.0
  description: |
    Comprehensive mental health platform API providing secure endpoints for:
    - User authentication and MFA management
    - Crisis assessment and intervention
    - Wellness data tracking and analytics
    - Payment processing and subscription management
    - Appointment scheduling and session management
    - Secure messaging and file management
    
    **Security Features:**
    - Rate limiting on all endpoints
    - PHI-compliant data encryption
    - Comprehensive audit logging
    - Role-based access control
    
    **Base URL:** `https://astral-core.app/api`
    
  contact:
    name: Astral Core API Support
    email: api-support@astral-core.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://astral-core.app/api
    description: Production server
  - url: https://staging.astral-core.app/api
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
security:
  - bearerAuth: []
  - sessionAuth: []
paths:
...
  /auth/mfa/verify:
    post:
      tags: [Authentication, MFA]
      summary: Verify MFA token
      description: Verify MFA token to complete authentication or setup
      operationId: verifyMFA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MFAVerifyRequest'
            examples:
              totp_verify:
                summary: TOTP token verification
                value:
                  method: "TOTP"
                  token: "123456"
              backup_code:
                summary: Backup code verification
                value:
                  method: "BACKUP_CODE"
                  token: "12345-67890"
      responses:
        '200':
          description: MFA verification successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MFAVerifyResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid MFA token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /appointments:
    get:
      tags: [Appointments]
      summary: Get user appointments
      description: |
        Retrieve user's appointments with filtering and pagination.
        Supports status filtering and pagination.
      operationId: getUserAppointments
      parameters:
        - name: status
          in: query
          description: Filter by appointment status
          schema:
            type: string
            enum: [SCHEDULED, CONFIRMED, IN_PROGRESS, COMPLETED, CANCELLED, NO_SHOW, RESCHEDULED]
            example: "SCHEDULED"
        - name: limit
          in: query
          description: Maximum number of appointments to return
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: offset
          in: query
          description: Number of appointments to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Appointments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentsResponse'
              examples:
                upcoming_appointments:
                  summary: Upcoming appointments
                  value:
                    success: true
                    data:
                      items:
                        - id: "cm3appt123"
                          scheduledAt: "2024-01-20T10:00:00Z"
                          duration: 60
                          type: "THERAPY_SESSION"
                          status: "SCHEDULED"
                          meetingUrl: "https://meet.astral-core.app/session/abc123"
                          notes: "Follow-up on anxiety management techniques"
                          reminderSent: true
                          therapist:
                            id: "cm3therapist456"
                            name: "Dr. Sarah Johnson"
                            email: "sarah.johnson@astral-core.app"
                      total: 5
                      page: 1
                      limit: 10
                      hasMore: false
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'

  # ==================== ADDITIONAL ENDPOINTS ====================
...
  /files:
    get:
      tags: [Files]
      summary: List user files
      description: Get a list of user's uploaded files with filtering and pagination
      operationId: getUserFiles
      parameters:
        - name: category
          in: query
          description: Filter by file category
          schema:
            type: string
            enum: [CONSENT_FORM, INSURANCE, MEDICAL_RECORD, SESSION_NOTE, ASSESSMENT, REPORT, OTHER]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Files retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/FileResponse/properties/file'
                      total:
                        type: integer
                      hasMore:
                        type: boolean
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /files/{id}:
    get:
      tags: [Files]
      summary: Download file
      description: Download a specific file by ID with proper authorization
      operationId: downloadFile
      parameters:
        - name: id
          in: path
          required: true
          description: File ID
          schema:
            type: string
        - name: inline
          in: query
          description: Whether to display inline or download
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: File downloaded successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Access denied to file
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

  # ==================== JOURNAL ENTRIES ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
...
    MFAVerifyRequest:
      type: object
      required:
        - method
        - token
      properties:
        method:
          type: string
          enum: [TOTP, SMS, EMAIL, BACKUP_CODE]
          description: MFA method used for verification
        token:
          type: string
          description: MFA token or backup code
          example: "123456"
...
    MFAVerifyResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        mfaEnabled:
          type: boolean
          description: Whether MFA is now enabled
...
    AppointmentsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Appointment'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            hasMore:
              type: boolean
    Appointment:
      type: object
      properties:
        id:
          type: string
        scheduledAt:
          type: string
          format: date-time
        duration:
          type: integer
        type:
          type: string
          enum: [INITIAL_CONSULTATION, THERAPY_SESSION, FOLLOW_UP, CRISIS_SESSION, GROUP_SESSION, ASSESSMENT]
        status:
          type: string
          enum: [SCHEDULED, CONFIRMED, IN_PROGRESS, COMPLETED, CANCELLED, NO_SHOW, RESCHEDULED]
        location:
          type: string
          nullable: true
        meetingUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        reminderSent:
          type: boolean
        therapist:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "An error occurred"
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              path:
                type: array
                items:
                  type: string
          description: Detailed validation errors (when applicable)

  # ==================== COMMON RESPONSES ====================
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              summary: Unauthorized access
              value:
                error: "Authentication required"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_failed:
              summary: Validation failed
              value:
                error: "Validation failed"
                details:
                  - code: "invalid_type"
                    message: "Expected string, received number"
                    path: ["email"]
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            not_found:
              summary: Resource not found
              value:
                error: "The requested resource was not found"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server_error:
              summary: Server error
              value:
                error: "An internal server error occurred"
...