# src/app/api/files/[id]/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { getServerSession } from 'next-auth';
import { authOptions } from '@/lib/auth/config';
import { fileUploadService } from '@/lib/services/file-upload-service';
import { HTTP_STATUS, ERROR_MESSAGES } from '@/lib/constants/index';

// GET /api/files/[id] - Download a file
export async function GET(request: NextRequest, context: { params: Promise<{ id: string }> }) {
  try {
    const session = await getServerSession(authOptions);
    if (!session?.user) {
      return NextResponse.json(
        { error: ERROR_MESSAGES.UNAUTHORIZED },
        { status: HTTP_STATUS.UNAUTHORIZED }
      );
    }

    const { id } = await context.params;
    const fileId = id;
    const { searchParams } = new URL(request.url);
    const download = searchParams.get('download') === 'true';
    const thumbnail = searchParams.get('thumbnail') === 'true';

    if (thumbnail) {
      // Handle thumbnail request
      return await handleThumbnail(fileId, session.user.id);
    }

    // Get file stream
    const { stream, filename, mimeType, size } = await fileUploadService.getFileStream(
      fileId,
      session.user.id
    );

    // Create response with proper headers
    const response = new NextResponse(stream as unknown as ReadableStream<Uint8Array>);

    response.headers.set('Content-Type', mimeType);
    response.headers.set('Content-Length', size.toString());

    if (download) {
      response.headers.set('Content-Disposition', `attachment; filename="${filename}"`);
    } else {
      response.headers.set('Content-Disposition', `inline; filename="${filename}"`);
    }

    // Security headers
    response.headers.set('X-Content-Type-Options', 'nosniff');
    response.headers.set('X-Frame-Options', 'DENY');

    // Cache control for files
    response.headers.set('Cache-Control', 'private, max-age=3600');

    return response;
  } catch (error: unknown) {
    console.error('Error serving file:', error);

    if (
      error instanceof Error &&
      (error.message === 'File not found' || error.message === 'Access denied')
    ) {
      return NextResponse.json({ error: error.message }, { status: HTTP_STATUS.NOT_FOUND });
    }

    return NextResponse.json(
      { error: ERROR_MESSAGES.SERVER_ERROR },
      { status: HTTP_STATUS.INTERNAL_SERVER_ERROR }
    );
  }
}

// DELETE /api/files/[id] - Delete a file
export async function DELETE(_request: NextRequest, context: { params: Promise<{ id: string }> }) {
  try {
    const session = await getServerSession(authOptions);
    if (!session?.user) {
      return NextResponse.json(
        { error: ERROR_MESSAGES.UNAUTHORIZED },
        { status: HTTP_STATUS.UNAUTHORIZED }
      );
    }

    const { id } = await context.params;
    await fileUploadService.deleteFile(id, session.user.id);

    return NextResponse.json({
      success: true,
      message: 'File deleted successfully'
    });
  } catch (error: unknown) {
    console.error('Error deleting file:', error);

    if (
      error instanceof Error &&
      (error.message === 'File not found' || error.message === 'Access denied')
    ) {
      return NextResponse.json({ error: error.message }, { status: HTTP_STATUS.NOT_FOUND });
    }

    return NextResponse.json(
      { error: ERROR_MESSAGES.SERVER_ERROR },
      { status: HTTP_STATUS.INTERNAL_SERVER_ERROR }
    );
  }
}

async function handleThumbnail(fileId: string, userId: string) {
  try {
    // This would serve the thumbnail file
    // For now, return a placeholder response
    console.log(`Thumbnail requested for file ${fileId} by user ${userId}`);
    return NextResponse.json(
      { error: 'Thumbnail service not implemented' },
      { status: HTTP_STATUS.NOT_IMPLEMENTED }
    );
  } catch (error: unknown) {
    console.error('Thumbnail error:', error);
    return NextResponse.json(
      { error: 'Thumbnail not available' },
      { status: HTTP_STATUS.NOT_FOUND }
    );
  }
}